WIP: change file name to DFRefinedRegistration and add documentation
